pyproject.toml
[project]
name = "x-reader"
version = "0.2.0"
description = "Universal content reader — fetch, normalize, and digest content from 7+ platforms"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Leo", email = "runes.leo@gmail.com"}
]
keywords = ["content-reader", "rss", "telegram", "bilibili", "xiaohongshu", "digest"]
dependencies = [
    "requests>=2.28",
    "feedparser>=6.0",
    "python-dotenv>=1.0",
    "loguru>=0.7",
    "idna>=3.6",  # For IDN/homograph attack protection in URL validation
]

[project.urls]
Homepage = "https://github.com/runesleo/x-reader"
Repository = "https://github.com/runesleo/x-reader"
Issues = "https://github.com/runesleo/x-reader/issues"

[project.optional-dependencies]
telegram = ["telethon>=1.34"]
mcp = ["mcp[cli]>=1.0"]
browser = ["playwright>=1.40"]
all = ["telethon>=1.34", "mcp[cli]>=1.0", "playwright>=1.40"]

[project.scripts]
x-reader = "x_reader.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["x_reader"]
